split regression tests into Dump vs. Script
